Move AthletesView into dedicated file
